+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [Unreleased] neptune-tensorboard 1.0.0.pre0
+
+### Changes
+- Update integration to work with latest `neptune` client and update the meta-data logging structure ([#51](https://github.com/neptune-ai/neptune-tensorboard/pull/51))
+
 ## neptune-tensorboard 0.5.3
 
 ### Changes